# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agorich Pharma API
  description: |
    Comprehensive API documentation for Agorich Pharma's hyperlocal pharma intelligence operating system.

    ## Overview
    This API provides endpoints for managing the pharmaceutical distribution ecosystem including:
    - AI-powered collection calls via VAPI
    - Inventory management and alerts
    - Invoice processing and approval workflows
    - Real-time metrics and analytics
    - Emergency controls and system management

    ## Authentication
    - **User Authentication**: Supabase Auth with JWT tokens (Cookie-based for web clients)
    - **Agent API Key**: `x-agent-api-key` header for backend-to-backend communication
    - **Admin API Key**: `x-admin-api-key` header for admin operations

    ## Rate Limiting
    - Default: 100 requests per minute per IP
    - AI endpoints: 20 requests per minute
    - Webhook endpoints: 200 requests per minute
  version: 1.0.0
  contact:
    name: Agorich Pharma Tech Team
    email: tech@agorichpharma.com
  license:
    name: Proprietary
servers:
  - url: https://api.agorichpharma.com
    description: Production server
  - url: https://staging-api.agorichpharma.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/admin/emergency/resume:
    post:
      tags:
        - Admin
      summary: Resume from emergency stop
      description: Deactivates the current emergency stop level.
      operationId: resumeFromEmergency
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                adminId:
                  type: string
      responses:
        '200':
          description: System resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmergencyResumeResponse'
...
  /api/admin/rollback/{actionId}:
    get:
      tags:
        - Admin
      summary: Get rollback action details
      description: Retrieves details of a specific rollback action.
      operationId: getRollbackAction
      security:
        - BearerAuth: []
      parameters:
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Action details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RollbackActionResponse'
  /api/admin/notifications:
    get:
      tags:
        - Admin
      summary: Get system notifications
      description: Retrieves system notifications for admin dashboard.
      operationId: getNotifications
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Notifications retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
    post:
      tags:
        - Admin
      summary: Create notification
      description: Creates a new system notification.
      operationId: createNotification
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationCreate'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-agent-api-key
      description: API key for agent-to-agent communication
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for administrative operations
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
    WebhookSecretAuth:
      type: apiKey
      in: header
      name: x-vapi-webhook-secret
      description: Webhook verification secret
  schemas:
...
    EmergencyResumeResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    RollbackActionResponse:
      type: object
      properties:
        success:
          type: boolean
        actionId:
          type: string
        actionType:
          type: string
        data:
          type: object
        timestamp:
          type: string
          format: date-time
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        created_at:
          type: string
          format: date-time
    NotificationsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        unreadCount:
          type: integer
    NotificationCreate:
      type: object
      required:
        - type
        - title
        - message
      properties:
        type:
          type: string
        title:
          type: string
        message:
          type: string
        user_id:
          type: string
...